* updated ac97 code with the one from Marcus' ich driver * merged existing quirks (reversed amp enable and ad1981b) testing is welcome!
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@25497 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/kernel/drivers/audio/ac97/auich/io.c
@@ -38,7 +38,7 @@
 extern pci_module_info  *pci;
 
 uint8
-auich_reg_read_8(device_config *config, int regno)
+auich_reg_read_8(device_config *config, uint8 regno)
 {
 	ASSERT(regno >= 0);
 	ASSERT(((config->type & TYPE_ICH4) != 0 && regno <= 255) || regno <= 63);
@@ -49,7 +49,7 @@ auich_reg_read_8(device_config *config, int regno)
 }
 
 uint16
-auich_reg_read_16(device_config *config, int regno)
+auich_reg_read_16(device_config *config, uint8 regno)
 {
 	ASSERT(regno >= 0);
 	ASSERT(((config->type & TYPE_ICH4) != 0 && regno <= 255) || regno <= 63);
@@ -60,7 +60,7 @@ auich_reg_read_16(device_config *config, int regno)
 }
 
 uint32
-auich_reg_read_32(device_config *config, int regno)
+auich_reg_read_32(device_config *config, uint8 regno)
 {
 	ASSERT(regno >= 0);
 	ASSERT(((config->type & TYPE_ICH4) != 0 && regno <= 255) || regno <= 63);
@@ -71,7 +71,7 @@ auich_reg_read_32(device_config *config, int regno)
 }
 
 void
-auich_reg_write_8(device_config *config, int regno, uint8 value)
+auich_reg_write_8(device_config *config, uint8 regno, uint8 value)
 {
 	ASSERT(regno >= 0);
 	ASSERT(((config->type & TYPE_ICH4) != 0 && regno <= 255) || regno <= 63);
@@ -82,7 +82,7 @@ auich_reg_write_8(device_config *config, int regno, uint8 value)
 }
 
 void
-auich_reg_write_16(device_config *config, int regno, uint16 value)
+auich_reg_write_16(device_config *config, uint8 regno, uint16 value)
 {
 	ASSERT(regno >= 0);
 	ASSERT(((config->type & TYPE_ICH4) != 0 && regno <= 255) || regno <= 63);
@@ -93,7 +93,7 @@ auich_reg_write_16(device_config *config, int regno, uint16 value)
 }
 
 void
-auich_reg_write_32(device_config *config, int regno, uint32 value)
+auich_reg_write_32(device_config *config, uint8 regno, uint32 value)
 {
 	ASSERT(regno >= 0);
 	ASSERT(((config->type & TYPE_ICH4) != 0 && regno <= 255) || regno <= 63);
@@ -118,7 +118,7 @@ auich_codec_wait(device_config *config)
 }
 
 uint16
-auich_codec_read(device_config *config, int regno)
+auich_codec_read(device_config *config, uint8 regno)
 {
 	ASSERT(regno >= 0);
 	ASSERT(((config->type & TYPE_ICH4) != 0 && regno <= 511) || regno <= 255);
@@ -134,7 +134,7 @@ auich_codec_read(device_config *config, int regno)
 }
 
 void
-auich_codec_write(device_config *config, int regno, uint16 value)
+auich_codec_write(device_config *config, uint8 regno, uint16 value)
 {
 	ASSERT(regno >= 0);
 	ASSERT(((config->type & TYPE_ICH4) != 0 && regno <= 511) || regno <= 255);
